Positive degree centrality computing works for unsigned graphs
All methods that compute PN centrality return the same values as signnet for table 5 and "Sampson Monastery" dataset. They still use signnet changes on formulas in the article. It is a temporary change used to tests.

If the functions that compute positive degree centrality are changed in the same way, they return the same values as signnet.

Fix errors in test files degree_centrality_test.py and degree_centrality_test.R.

Updated files:
    degree_centrality.py
    degree_centrality_test.py
    degree_centrality_test.R