Modified files to match Nejat outputs file columns
Minor modifications done in the code : I replaced "imbalance" by
"imbalance_count" whenever it was necessery to match with the Nejat outputs
file columns name.
Also, there is a new file that I'll use to collect and save predicted
data into csv files. Right now there is nothing in it, I'll update it
when I find the best way to do it.

Added file :
	src/collect/collect_predicted_values.py

Modified files :
	src/collect_outputs.py
	src/consts.py
	src/main.py