Small fix in generate_boxpot_clean method
I modifed my code at the last commit, but I forgot something.
I coded in "brute force" an interval of 10, but I changed it to a
variable at the beginning of the method. But, I forgot to modify another
code line (line 201 of this commit) to make a completely modular method.
This omission is now fixed.

Modified file :
	src/collect/collect_graphics.py