Feature ablation updated (still unusable globally)
I tried to use a regression task to see if the "same score everytime"
issue is still here, but it worked with this task, and I don't know why.
I'll try to refactor the module to use any type of predictor, and maybe
the issue will be resolved by itself.

modified files :
	src/main.py
	src/prediction/feature_ablation.py